pyproject.toml
[project]
name = "log-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "colorlog>=6.10.1",
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]
